Integrate Igniter for automated project setup
- Add {:igniter, "~> 0.7", optional: true} to deps
- Rewrite mix ex_tauri.install with Igniter support:
  - Automatically adds ShutdownManager to supervision tree
  - Automatically adds :desktop Burrito release to mix.exs
  - Falls back to manual instructions when Igniter not available
- Extract shared helpers to ExTauri.Install.Helpers module
- Simplify E2E test: no more fragile regex for supervision tree
  and mix.exs project config — Igniter handles it via AST patching
- E2E test now adds igniter as a dep alongside ex_tauri
