+++ mix.exs
@@ -26,6 +26,7 @@ defmodule ExTauri.MixProject do
     [
       {:ex_doc, ">= 0.0.0", only: :docs},
       {:burrito, "~> 1.5"},
+      {:igniter, "~> 0.7", optional: true},
       {:jason, "~> 1.4.0"}
     ]
   end
